refactor(protectedroutewithprops): shifts into own file and migrate to ts
this is required because otherwise, storybook complains about workspace not being initialised.
