+++ src/layouts/ResourceRoom/ResourceRoom.tsx
@@ -36,7 +36,7 @@ import {
   DirectorySettingsScreen,
 } from "layouts/screens"
 
-import { ProtectedRouteWithProps } from "routing/RouteSelector"
+import { ProtectedRouteWithProps } from "routing/ProtectedRouteWithProps"
 
 import { useErrorToast } from "utils/toasts"
